rename playback panel folder
